Updated readme and fixed vulnerabilities
Changed the location of where to look for the wallet in the client applications. As of the latest version of the plugin, the identity for the local_fabric admin is stored in $HOME_DIR/.fabric-vscode/local_fabric_wallet. This admin identity is also automatically created for you so there is no need to run the addToWallet.js script.

Also, fixed vulnerabilities found in the js-yaml and mqtt-packet modules.

Updated readme with more clear instructions and fixed spelling errors.

Signed-off-by: Oliver Rodriguez <odrodrig@us.ibm.com>