Changed where to find the local wallet and removed addToWallet.js step
In a recent update to the IBM Blockchain Platform extension the local fabric wallet was moved. This broke the applications becuase the apps are looking for the local wallet to pull the admin identity. This admin identity is automatically added to the local wallet by the extension as part of network creation so there is no need for the addToWallet.js file. The steps regarding adding the identity to the wallet manually have been deleted from the readme.

Signed-off-by: Oliver Rodriguez <odrodrig@us.ibm.com>